whois: makes a bit more sense now.

+++ userbot/modules/whois.py
@@ -20,30 +20,31 @@ from userbot.events import register, errors_handler
 @errors_handler
 async def who(event):
     """ For .whois command, get info about a user. """
-    await event.edit(
-        "`Sit tight while I steal some data from Mark Zuckerburg...`")
-
     if not os.path.isdir(TEMP_DOWNLOAD_DIRECTORY):
         os.makedirs(TEMP_DOWNLOAD_DIRECTORY)
-
     replied_user = await get_user(event)
-    photo, caption = await fetch_info(replied_user, event)
-    message_id_to_reply = event.message.reply_to_msg_id
-    if not message_id_to_reply:
-        message_id_to_reply = None
-    try:
-        await event.client.send_file(event.chat_id,
-                                     photo,
-                                     caption=caption,
-                                     link_preview=False,
-                                     force_document=False,
-                                     reply_to=message_id_to_reply,
-                                     parse_mode="html")
-        if not photo.startswith("http"):
-            os.remove(photo)
-        await event.delete()
-    except TypeError:
-        await event.edit(caption, parse_mode="html")
+    if not replied_user:
+        return
+    else:
+        await event.edit(
+            "`Sit tight while I steal some data from Mark Zuckerburg...`")
+        photo, caption = await fetch_info(replied_user, event)
+        message_id_to_reply = event.message.reply_to_msg_id
+        if not message_id_to_reply:
+            message_id_to_reply = None
+        try:
+            await event.client.send_file(event.chat_id,
+                                         photo,
+                                         caption=caption,
+                                         link_preview=False,
+                                         force_document=False,
+                                         reply_to=message_id_to_reply,
+                                         parse_mode="html")
+            if not photo.startswith("http"):
+                os.remove(photo)
+            await event.delete()
+        except TypeError:
+            await event.edit(caption, parse_mode="html")
 
 
 async def get_user(event):
